Dumps the current implementation as one commit
Commits the a blob of code that has all the implementation so far. The
implementation covers most of the core functionality, but it's not been
tested yet. This is to preserve whatever is available at the moment and
be able to track the future changes.

Related Issue(s): N/A